chore: Add confirm and update-alternatives.

+++ third-party.fish
@@ -10,24 +10,21 @@ if sudo which pip >/dev/null && confirm 'Do you want to install system-level sof
     sudo pip install -r pkglist/requirements-global.txt
 end
 
-if type -q pipx
-end
-
-if sudo which pipx >/dev/null
-    for pkg in (cat ./pkglist/pipx-global.txt | grep -v '^#')
+if sudo which pipx >/dev/null && confirm 'Do you want to install system-level software using pipx?'
+    for pkg in (cat ./pkglist/pipx-system.txt | grep -v '^#')
         sudo fish -c "pipx install $pkg"
     end
-    for inject in (cat ./pkglist/pipx-inject-global.txt | grep -v '^#')
+    for inject in (cat ./pkglist/pipx-inject-system.txt | grep -v '^#')
         sudo fish -c "pipx inject $inject"
     end
 end
 
-if type -q go
+if type -q go && confirm 'Do you want to install some software using go?'
     go install github.com/davecheney/httpstat@latest
     go install github.com/shurcooL/markdownfmt@latest
 end
 
-if type -q fnm
+if type -q fnm && confirm 'Do you want to install some software using fnm?'
     fnm install --lts
 end
 
@@ -36,18 +33,18 @@ if type -q micromamba && confirm 'Do you want to install some software using mic
     micromamba run -n labelme pip install -U labelme
 end
 
-if sudo which yarn >/dev/null
+if sudo which yarn >/dev/null && confirm 'Do you want to install some software using yarn?'
     sudo yarn global add (cat ./pkglist/yarn-global.txt | grep -v '^#')
-else if type -q yarn
+else if type -q yarn && confirm 'Do you want to install some software using yarn?'
     yarn global add (cat ./pkglist/yarn-global.txt | grep -v '^#')
 end
 
-if type -q wine
+if type -q wine && confirm 'Do you want to install some software using wine?'
     aria2c --conf-path=./config/aria2/aria2.conf \
-        'https://mirrors.cernet.edu.cn/winehq/wine/wine-mono/8.0.0/wine-mono-8.0.0-x86.msi' \
+        'https://unicom.mirrors.ustc.edu.cn/wine/wine/wine-mono/9.1.0/wine-mono-9.1.0-x86.msi' \
         'https://mirrors.cernet.edu.cn/winehq/wine/wine-gecko/2.47.4/wine-gecko-2.47.4-x86_64.msi' \
         'https://mirrors.cernet.edu.cn/winehq/wine/wine-gecko/2.47.4/wine-gecko-2.47.4-x86.msi'
-    wine msiexec /i ./wine-mono-8.0.0-x86.msi
+    wine msiexec /i ./wine-mono-9.1.0-x86.msi
     wine msiexec /i ./wine-gecko-2.47.4-x86_64.msi
     wine msiexec /i ./wine-gecko-2.47.4-x86.msi
 end
